Updated with changes from OmniGroup svn [292307:292812]
Configurations:
- Bump the time for -warn-long-function-bodies from 100ms to 125ms. There are a
  lot of trivial functions that just barely miss 100ms, so let's concentrate on
  the more complicated stuff if we are going to fix something first.

Tools/SVNVersion:
- List the output SVNVersion.h file for each of the per-platform shell script
  build phases.

OmniBase:
- On iOS, adding a block observer for NSBundleDidLoadNotification causes
  a hang in CoreAnimation when initializing a UIWebView (predictably) or
  even when just rotating the screen (and taking a snapshot of the
  existing views).

  We don't need this notification on iOS (where we don't load bundles
  dynamically anyway and where we'd never even ported OBPostLoader
  functionality in the past), so I'm disabling the observation code
  there.

OmniFoundation:
- Add OFUTIDescription.
- Modernize some code in OFCompletionMatch.
- Fixed undefined behavior noticed by UBSan.

OmniDataObjects:
- Added convenience extensions on ODOEditingContext to allow fetching of an
  object by objectID or primaryKey/entityName pair that infers the type from
  the calling context to avoid casting at the call site.

OmniUI/iPad:
- More themeing work in inspectors.
- Add support for displaying different content in a view controller when a
  feature is disabled.
- OUIPasswordAlert no longer drops its provided protection space on the floor.
- Reworked hideSidebar(atLocation:) so that it does the correct thing for both
  overlaid and embedded sidebars.
- Handle all of our special scheme URLs being loaded by the web view,
  not just those triggered by a "link activated" navigation.

OWF:
- Fixed a crash due to sending a message to an OFWeakReference that was
  meant to go to the object it holds.

OmniAppKit:
- When computing a composite color, if the top most color is opaque just use it.
- When converting a color to a destination color space, if the color is already
  in that color space, just use it.

+++ Frameworks/OmniFoundation/XML/OFXMLIdentifier.m
@@ -116,7 +116,7 @@ NSString *OFXMLCreateIDFromData(NSData *data)
     while (inputIndex < inputSize || bitsInBuffer > 0) {
         if (bitsInBuffer <= 24 && inputIndex < inputSize) {
             // Room in buffer for 8 more bits
-            buffer |= input[inputIndex] << bitsInBuffer;
+            buffer |= ((uint32_t)input[inputIndex]) << bitsInBuffer;
             inputIndex++;
             bitsInBuffer += 8;
         }
